refactor(reborn): split test_support.rs into per-seam-family submodules
Addresses review: the file crossed 1000 lines with this PR's E-DURABLE/E-SKILL
additions, and kept growing as a flat catch-all of independent seam families
(E-PROFILE, E-PROJ, OAuth/product-auth, secrets/local-dev-boot, budget
gateway, E-DURABLE, E-SKILL). Converted to a directory module
(`test_support/`) mirroring the existing `runtime/` split pattern in this
same crate: one file per seam family, `mod.rs` is a thin re-export layer
listing the full public surface at a glance. Every doc comment preserved
verbatim; module-file boundaries only, no behavior change.

Also: fixed a stale E-SKILL doc comment in harness.rs that still named a
nonexistent `RebornIntegrationGroup::skill_tools` ctor with tenant/user
scoping — the real ctor is `skill_activation_tools`, seeding a system-scoped
skill.

Verified: build clean (plain + --features test-support + --features libsql +
full workspace --features libsql check); all 6 affected test binaries green
(reborn_integration_{cancel,durable,skill_activate}, reborn_group_approvals,
reborn_integration_{greeting,secret_injection}); clippy --all-features clean.
