+++ crates/ironclaw_reborn_composition/src/test_support/oauth_product_auth.rs
@@ -1,209 +1,13 @@
-//! Test-only helpers for the Reborn integration-test framework and budget E2E tests.
+//! OAuth / product-auth test bundles (Slices 7-8).
 //!
-//! Gated behind the `test-support` feature so production builds never pay the cost
-//! of the mock gateway / introspection accessors. The module covers three areas:
-//!
-//! 1. **Budget / mock-gateway helpers** — [`BudgetTestGateway`], [`FailingTestGateway`],
-//!    [`ScriptedReply`] — scripted model responses with configurable token counts for
-//!    `RebornRuntimeInput::with_model_gateway_override` tests.
-//! 2. **OAuth / product-auth test bundles** — [`ScriptedOAuthTokenEgress`],
-//!    [`OAuthProductAuthTestBundle`], `build_oauth_product_auth_for_test`,
-//!    `build_google_oauth_product_auth_for_test` — real store / real client / scripted
-//!    HTTP egress for OAuth connect, refresh, and error-path tests.
-//! 3. **Reborn integration-test framework accessors** — `build_local_dev_approval_gate_evidence_for_test`,
-//!    `build_default_local_dev_database_roots_for_test`, `mount_local_dev_database_roots_for_test`,
-//!    `build_local_dev_secret_store_for_test` — mirror the production local-dev boot
-//!    sequence so the integration-test harness (`tests/support/reborn/`) drives the
-//!    real local-dev composition paths without duplicating the wiring logic.
+//! [`ScriptedOAuthTokenEgress`], [`OAuthProductAuthTestBundle`],
+//! `build_oauth_product_auth_for_test`, `build_google_oauth_product_auth_for_test`
+//! — real store / real client / scripted HTTP egress for OAuth connect,
+//! refresh, and error-path tests.
 
 use std::sync::{Arc, Mutex};
 
 use async_trait::async_trait;
-use ironclaw_host_api::ThreadId;
-use ironclaw_loop_support::{
-    HostManagedModelError, HostManagedModelErrorKind, HostManagedModelGateway,
-    HostManagedModelRequest, HostManagedModelResponse,
-};
-use ironclaw_turns::{
-    TurnRunId, TurnStatus,
-    run_profile::{LoopCapabilityPort, LoopModelUsage},
-};
-
-/// Filename of the local-dev libSQL database within the per-user root
-/// directory. Value is derived from the production factory constant so
-/// there is one owner of the string; tests access it through this
-/// test-support surface.
-pub const LOCAL_DEV_DB_FILENAME: &str = crate::factory::LOCAL_DEV_DB_FILENAME;
-
-use crate::runtime::{AssistantReply, ConversationId};
-
-/// Build a terminal/no-text assistant reply for CLI and product-surface tests.
-///
-/// Kept behind `test-support` so downstream crates can exercise presentation
-/// paths without depending directly on lower-level turn/thread crates.
-pub fn assistant_reply_without_text_for_test(
-    status: TurnStatus,
-    failure_category: Option<&str>,
-) -> AssistantReply {
-    AssistantReply {
-        conversation: ConversationId(
-            ThreadId::new("test-assistant-reply").expect("static test thread id"), // safety: static test helper id is a valid thread id literal.
-        ),
-        run_id: TurnRunId::new(),
-        status,
-        failure_category: failure_category.map(str::to_owned),
-        text: None,
-    }
-}
-
-/// One scripted reply from the mock LLM.
-///
-/// `usage` is forwarded into [`HostManagedModelResponse::usage`] so the
-/// budget accountant reconciles against real provider numbers, not the
-/// reservation estimate.
-#[derive(Debug, Clone)]
-pub struct ScriptedReply {
-    pub text: String,
-    pub input_tokens: u32,
-    pub output_tokens: u32,
-}
-
-impl ScriptedReply {
-    pub fn new(text: impl Into<String>, input_tokens: u32, output_tokens: u32) -> Self {
-        Self {
-            text: text.into(),
-            input_tokens,
-            output_tokens,
-        }
-    }
-
-    fn into_response(self) -> HostManagedModelResponse {
-        HostManagedModelResponse::assistant_reply(self.text).with_usage(LoopModelUsage {
-            input_tokens: self.input_tokens,
-            output_tokens: self.output_tokens,
-        })
-    }
-}
-
-/// Mock [`HostManagedModelGateway`] that returns scripted assistant
-/// replies with configurable token usage.
-///
-/// Replies are consumed in FIFO order. When the script runs out the
-/// gateway falls back to a sentinel reply with zero tokens — tests that
-/// drive multiple turns should pre-load the matching number of
-/// [`ScriptedReply`] entries.
-///
-/// Every `stream_model` call is recorded so tests can assert the call
-/// count after the run completes.
-#[derive(Debug, Default)]
-pub struct BudgetTestGateway {
-    replies: Mutex<Vec<ScriptedReply>>,
-    fallback: Option<ScriptedReply>,
-    calls: Mutex<Vec<HostManagedModelRequest>>,
-}
-
-impl BudgetTestGateway {
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Single-reply convenience: every model call returns the same
-    /// assistant text with the given token counts.
-    pub fn with_constant(text: impl Into<String>, input_tokens: u32, output_tokens: u32) -> Self {
-        Self {
-            replies: Mutex::new(Vec::new()),
-            fallback: Some(ScriptedReply::new(text, input_tokens, output_tokens)),
-            calls: Mutex::new(Vec::new()),
-        }
-    }
-
-    /// Push one scripted reply. Replies are consumed in FIFO order.
-    pub fn push(&self, reply: ScriptedReply) {
-        self.replies
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-            .push(reply);
-    }
-
-    /// Number of `stream_model` calls observed so far.
-    pub fn call_count(&self) -> usize {
-        self.calls
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-            .len()
-    }
-
-    fn next_reply(&self) -> ScriptedReply {
-        let mut script = self
-            .replies
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner);
-        if script.is_empty() {
-            return self
-                .fallback
-                .clone()
-                .unwrap_or_else(|| ScriptedReply::new("budget test fallback reply", 0, 0));
-        }
-        script.remove(0)
-    }
-}
-
-#[async_trait]
-impl HostManagedModelGateway for BudgetTestGateway {
-    async fn stream_model(
-        &self,
-        request: HostManagedModelRequest,
-    ) -> Result<HostManagedModelResponse, HostManagedModelError> {
-        self.calls
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-            .push(request);
-        Ok(self.next_reply().into_response())
-    }
-
-    async fn stream_model_with_capabilities(
-        &self,
-        request: HostManagedModelRequest,
-        _capabilities: Arc<dyn LoopCapabilityPort>,
-    ) -> Result<HostManagedModelResponse, HostManagedModelError> {
-        // The budget tests don't need capability dispatch — fall through
-        // to the plain stream path. If a future test needs tool calls,
-        // extend this with a separate scripted-tool-call queue.
-        self.calls
-            .lock()
-            .unwrap_or_else(std::sync::PoisonError::into_inner)
-            .push(request);
-        Ok(self.next_reply().into_response())
-    }
-}
-
-/// Mock gateway that always fails with the given error kind. Useful for
-/// driving the cancellation / provider-error paths in budget tests
-/// without depending on tokio cancel semantics.
-#[derive(Debug)]
-pub struct FailingTestGateway {
-    pub kind: HostManagedModelErrorKind,
-    pub summary: String,
-}
-
-impl FailingTestGateway {
-    pub fn new(kind: HostManagedModelErrorKind, summary: impl Into<String>) -> Self {
-        Self {
-            kind,
-            summary: summary.into(),
-        }
-    }
-}
-
-#[async_trait]
-impl HostManagedModelGateway for FailingTestGateway {
-    async fn stream_model(
-        &self,
-        _request: HostManagedModelRequest,
-    ) -> Result<HostManagedModelResponse, HostManagedModelError> {
-        Err(HostManagedModelError::safe(self.kind, self.summary.clone()))
-    }
-}
 
 // ─── Slice 7: OAuth connect-flow test support ─────────────────────────────────
 //
@@ -750,253 +554,4 @@ pub fn build_google_oauth_product_auth_for_test() -> OAuthProductAuthTestBundle 
     );
 
     OAuthProductAuthTestBundle { services, egress }
-}
-
-/// Test-only accessor mirroring the full local-dev database-roots boot path
-/// (`build_local_dev_root_filesystem` → `build_default_local_dev_database_roots`).
-///
-/// Constructs the durable database backend and mounts it across the
-/// control-plane roots (`/tenants`, `/memory`, `/events`) of `composite`,
-/// selecting the backend by compile-time feature:
-/// - With `libsql`: opens `root/reborn-local-dev.db`, runs migrations, mounts.
-/// - Without a durable backend feature: mounts an in-memory backend.
-///
-/// Called by the Reborn integration-test framework's `StorageMode::LibSql`
-/// builder arm (`tests/support/reborn/builder.rs:build_storage_composite`) so
-/// the 4-step libSQL setup sequence lives once (production call site:
-/// `build_local_dev_root_filesystem` → `build_default_local_dev_database_roots`).
-/// For tests only — gated behind `test-support`, ships zero bytes in production.
-#[cfg(feature = "test-support")]
-pub async fn build_default_local_dev_database_roots_for_test(
-    root: &std::path::Path,
-    composite: &mut ironclaw_filesystem::CompositeRootFilesystem,
-) -> Result<(), crate::RebornBuildError> {
-    crate::factory::mount_default_local_dev_database_roots(root, composite).await
-}
-
-/// Test-only accessor mirroring the production local-dev boot path
-/// (`build_local_dev_root_filesystem` → `mount_local_dev_database_roots`).
-///
-/// Mounts `database` across the control-plane roots (`/tenants`, `/memory`,
-/// `/events`) of `root` exactly as the libSQL local-dev boot path does, so
-/// downstream integration tests (the Reborn integration-test framework in
-/// `tests/support/reborn/`) construct one real `LibSqlRootFilesystem` over a
-/// composite without a second copy of the mount wiring (design spec §3.2).
-/// For tests only — gated behind `test-support`, so it ships zero bytes in
-/// production binaries.
-#[cfg(feature = "test-support")]
-pub fn mount_local_dev_database_roots_for_test<F>(
-    root: &mut ironclaw_filesystem::CompositeRootFilesystem,
-    database: std::sync::Arc<F>,
-) -> Result<(), crate::RebornBuildError>
-where
-    F: ironclaw_filesystem::RootFilesystem + 'static,
-{
-    crate::factory::mount_local_dev_database_roots(root, database)
-}
-
-/// Test-only entry point for building a local-dev
-/// [`ironclaw_secrets::FilesystemSecretStore`] without going through the full
-/// Reborn runtime assembly.
-///
-/// Mirrors the production wiring in `build_local_runtime` where
-/// `build_local_dev_secret_store` is called with the scoped filesystem and a
-/// master key resolved from the environment or the root directory's cached key
-/// file. Tests that need a real `FilesystemSecretStore` — for example, to
-/// verify `put` + `lease_once` + `consume` round-trips against an in-process
-/// backend — can call this instead of wiring a full runtime.
-///
-/// The master key is resolved exactly as production does: from the
-/// `SECRETS_MASTER_KEY` env var when set, otherwise from (or generating to)
-/// the `.reborn-local-dev-secrets-master-key` file under `root`. Using the
-/// same `root` across two calls therefore yields the same key, so a second
-/// `FilesystemSecretStore` over the same scoped filesystem can consume a
-/// secret written by the first. For tests only — zero bytes shipped in
-/// production builds.
-#[cfg(any(feature = "libsql", feature = "postgres"))]
-pub fn build_local_dev_secret_store_for_test<F>(
-    root: &std::path::Path,
-    scoped: std::sync::Arc<ironclaw_filesystem::ScopedFilesystem<F>>,
-) -> Result<std::sync::Arc<ironclaw_secrets::FilesystemSecretStore<F>>, crate::RebornBuildError>
-where
-    F: ironclaw_filesystem::RootFilesystem + 'static,
-{
-    crate::factory::build_local_dev_secret_store(root, scoped, None)
-}
-
-/// Mirrors the production approval-gate evidence wiring done by
-/// `build_local_runtime` (runtime.rs ~line 2799) — returns the REAL
-/// `LocalDevApprovalGateEvidence` so the gate-evidence lookup logic
-/// (the `gate:approval-` prefix parse + `ApprovalStatus::Pending` check)
-/// never drifts from production. Tests only.
-///
-/// Wired by the Reborn integration-test harness's one-runtime group assembly
-/// (`tests/support/reborn/group.rs`'s `into_group`, which builds the group's
-/// single planned runtime via `build_default_planned_runtime`) so a
-/// `BlockedApproval` run is verified against the persisted `Pending` approval
-/// request at loop exit and genuinely pauses — mirrors the production
-/// `runtime.rs` path with the real type, never a hand-mirrored copy.
-#[cfg(feature = "test-support")]
-pub fn build_local_dev_approval_gate_evidence_for_test(
-    approval_requests: std::sync::Arc<dyn ironclaw_run_state::ApprovalRequestStore>,
-) -> std::sync::Arc<dyn ironclaw_reborn::loop_exit_applier::ApprovalGateEvidenceStore> {
-    crate::runtime::build_local_dev_approval_gate_evidence_for_test(approval_requests)
-}
-
-/// Capability id of the local-dev synthetic `project_create` capability
-/// (E-PROJ seam). Single owner is the production constant in
-/// `runtime::local_dev::project_create`; the harness references this so its
-/// `project_tools()` constructor and assertions never hardcode the string.
-#[cfg(feature = "test-support")]
-pub const PROJECT_CREATE_CAPABILITY_ID: &str = crate::runtime::PROJECT_CREATE_CAPABILITY_ID;
-
-/// Test-support entry point (E-DURABLE seam): reopen a fresh, independent
-/// extension-installation store at an existing local-dev `storage_root`. Lets
-/// the integration harness prove capability-produced durable state survives a
-/// reopen, paralleling `assert_reply_persists_after_reopen`. Delegates to the
-/// production filesystem mounts + install-store load in `factory` so the reopen
-/// path never drifts from `build_reborn_services`. Tests only.
-#[cfg(feature = "test-support")]
-pub async fn open_local_dev_extension_installation_store_for_test(
-    storage_root: &std::path::Path,
-) -> Result<
-    std::sync::Arc<dyn ironclaw_extensions::ExtensionInstallationStore>,
-    crate::RebornBuildError,
-> {
-    crate::factory::open_local_dev_extension_installation_store_for_test(storage_root).await
-}
-
-/// Test-support entry point for the `project_create` synthetic-capability wrap
-/// (E-PROJ seam). Lets the integration-test harness inject the synthetic
-/// `project_create` capability onto its host-runtime capability port via the
-/// real production wrap (`wrap_local_dev_synthetic_capabilities` +
-/// `project_create_capability`), so the dispatch path never drifts from
-/// production.
-#[cfg(feature = "test-support")]
-pub fn wrap_project_create_capability_for_test(
-    inner: std::sync::Arc<dyn ironclaw_turns::run_profile::LoopCapabilityPort>,
-    project_service: std::sync::Arc<dyn ironclaw_product_workflow::ProjectService>,
-    fallback_user_id: ironclaw_host_api::UserId,
-    run_context: ironclaw_turns::run_profile::LoopRunContext,
-    input_resolver: std::sync::Arc<dyn ironclaw_loop_support::LoopCapabilityInputResolver>,
-    result_writer: std::sync::Arc<dyn ironclaw_loop_support::LoopCapabilityResultWriter>,
-) -> Result<
-    std::sync::Arc<dyn ironclaw_turns::run_profile::LoopCapabilityPort>,
-    ironclaw_turns::run_profile::AgentLoopHostError,
-> {
-    crate::runtime::wrap_project_create_capability_for_test(
-        inner,
-        project_service,
-        fallback_user_id,
-        run_context,
-        input_resolver,
-        result_writer,
-    )
-}
-
-/// Capability id of the local-dev synthetic `skill_activate` capability
-/// (E-SKILL seam). Single owner is the production constant in
-/// `runtime::local_dev::skill_activation`; mirrors `PROJECT_CREATE_CAPABILITY_ID`.
-#[cfg(feature = "test-support")]
-pub const SKILL_ACTIVATE_CAPABILITY_ID: &str = crate::runtime::SKILL_ACTIVATE_CAPABILITY_ID;
-
-/// Opaque handle (E-SKILL seam) carrying the built local-dev skill context
-/// source. Hides the crate-private `LocalDevSelectableSkillContextSource` from
-/// the integration-test crate, which cannot name it. Exposes the
-/// `HostSkillContextSource` for runtime wiring; the activation source travels
-/// inside for [`wrap_skill_activation_capability_for_test`]. Tests only.
-#[cfg(feature = "test-support")]
-pub struct SkillActivationTestSource {
-    source: std::sync::Arc<dyn ironclaw_loop_support::HostSkillContextSource>,
-    activation_source: std::sync::Arc<crate::runtime::LocalDevSelectableSkillContextSource>,
-}
-
-#[cfg(feature = "test-support")]
-impl SkillActivationTestSource {
-    /// The `HostSkillContextSource` to wire as the runtime's
-    /// `skill_context_source` (`into_group`, E-SKILL) so activated-skill
-    /// instructions inject into the model request.
-    pub fn context_source(
-        &self,
-    ) -> std::sync::Arc<dyn ironclaw_loop_support::HostSkillContextSource> {
-        self.source.clone()
-    }
-}
-
-/// Build the local-dev skill context source (`HostSkillContextSource` for
-/// prompt injection plus the activation source backing `skill_activate`) over
-/// the runtime's skill filesystem, mirroring production `build_reborn_runtime`
-/// wiring (runtime.rs ~line 2875). Returns `None` when no local runtime is
-/// composed. Mirrors `build_user_profile_source_for_test` (E-SKILL seam).
-/// Tests only.
-#[cfg(feature = "test-support")]
-pub fn build_local_dev_skill_context_source_for_test(
-    services: &crate::RebornServices,
-    tenant_id: &ironclaw_host_api::TenantId,
-    regex_skill_activation_enabled: bool,
-) -> Option<SkillActivationTestSource> {
-    let local_runtime = services.local_runtime.as_ref()?;
-    // `None` means "no local runtime composed" (a legitimate backend shape,
-    // handled by the `?` above). A build *error* is a genuine misconfiguration
-    // of the local-dev skill filesystem, so surface it loudly rather than
-    // masking it as an un-wired skill source (which would fail a skill test at
-    // a confusing, far-removed assertion instead of here). Test-only code, so a
-    // panic is the right failure mode.
-    let (source, activation_source) =
-        crate::runtime::local_dev_filesystem_skill_context_source_for_test(
-            local_runtime,
-            tenant_id,
-            regex_skill_activation_enabled,
-        )
-        .expect("build local-dev skill context source for test");
-    Some(SkillActivationTestSource {
-        source,
-        activation_source,
-    })
-}
-
-/// Test-support entry point (E-SKILL seam): inject the `skill_activate`
-/// synthetic capability onto `inner`, using the source built by
-/// [`build_local_dev_skill_context_source_for_test`]. Mirrors
-/// `wrap_project_create_capability_for_test`; the dispatch path never drifts
-/// from production. Tests only.
-#[cfg(feature = "test-support")]
-pub fn wrap_skill_activation_capability_for_test(
-    inner: std::sync::Arc<dyn ironclaw_turns::run_profile::LoopCapabilityPort>,
-    source: &SkillActivationTestSource,
-    run_context: ironclaw_turns::run_profile::LoopRunContext,
-    input_resolver: std::sync::Arc<dyn ironclaw_loop_support::LoopCapabilityInputResolver>,
-    result_writer: std::sync::Arc<dyn ironclaw_loop_support::LoopCapabilityResultWriter>,
-) -> Result<
-    std::sync::Arc<dyn ironclaw_turns::run_profile::LoopCapabilityPort>,
-    ironclaw_turns::run_profile::AgentLoopHostError,
-> {
-    crate::runtime::wrap_skill_activation_capability_for_test(
-        inner,
-        source.activation_source.clone(),
-        run_context,
-        input_resolver,
-        result_writer,
-    )
-}
-
-/// Build the `HostUserProfileSource` the Reborn integration harness wires into
-/// the group's single planned runtime in `into_group`
-/// (`tests/support/reborn/group.rs`, E-PROFILE seam).
-///
-/// Reuses the production `MemoryBackedUserProfileSourceAdapter` (the single
-/// orphan-rule wrapper around `MemoryBackedUserProfileSource`) so the test path
-/// never drifts from production wiring (runtime.rs ~line 3167). When `filesystem`
-/// is `Some`, profile reads come from that raw local-dev memory filesystem;
-/// `None` (non-HostRuntime backends) falls back to `EmptyUserProfileSource`.
-#[cfg(feature = "test-support")]
-pub fn build_user_profile_source_for_test(
-    filesystem: Option<std::sync::Arc<dyn ironclaw_filesystem::RootFilesystem>>,
-) -> std::sync::Arc<dyn ironclaw_loop_support::HostUserProfileSource> {
-    match filesystem {
-        Some(fs) => std::sync::Arc::new(crate::runtime::MemoryBackedUserProfileSourceAdapter(
-            ironclaw_host_runtime::MemoryBackedUserProfileSource::new(fs),
-        )),
-        None => std::sync::Arc::new(ironclaw_loop_support::EmptyUserProfileSource),
-    }
 }
